Fixed note calories and portion data type (integer -> double)

+++ src/main/kotlin/com/oreyo/data/table/NoteTable.kt
@@ -9,11 +9,11 @@ object NoteTable: Table() {
 	val uid = reference("uid", UserTable.uid)
 	val noteId = varchar("note_id", 128)
 	val category = varchar("category", 48)
-	val calories = integer("calories")
+	val calories = double("calories")
 	val date = varchar("date", 48)
 	val food = varchar("food", 128)
 	val information = varchar("information", 256)
-	val portion = integer("portion")
+	val portion = double("portion")
 	val time = varchar("time", 24)
 	
 	override val primaryKey: PrimaryKey = PrimaryKey(noteId)
